QuieroAprender/QuieroAprender/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-ckeditor==6.7.2
django-js-asset==2.2.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
